Remove prepublish
For insane reasons, NPM crew chose to make `prepublish` trigger both on `npm publish` and on (some cases of) `npm install`, rather than having a special hook for local installs. Since we want CI to build things once and then test those same artifacts throughout the the whole QA pipeline, we do not want to rebuild the project right before deploy, and thus do not want the prepublish hook.

I was going to just sed out prepublish in CI, until I realized that prepublish only triggers if you manually `git checkout <repo> && npm install`, it does not trigger if you do `npm install <repo>`, which seems to be the idiomatic pattern. Hence, prepublish is inconsistent and gets in the way of CI. People that manually `git checkout` can simply `npm build`.

We will need to address `npm install <repo>` somehow - and we should find a way to do it that means we don't need to keep a built version of the library checked into the code base.