Add support for TLS encrypted connections.
This was a bit tricky, several things play into making this as
complicated as it turned out:

- TOFU (Trust on First Use) as we implement it only works on NodeJS,
  and only since https://github.com/nodejs/node/commit/345c40b6
- Browsers do not support specifying which certificates to trust,
  so CA list configuration only works on NodeJS
- The location of `.neo4j/known_hosts`, used by TOFU, differs between windows
  and linux.

However, tests are now green, all rejoice. For a modern NodeJS
deployment, the driver now defaults to using encryption and uses TOFU
for establishing trust, meaning users will get a low-hassle encrypted
setup by default.

If users are running in a web browser, or running a several-years-old
version of NodeJS, the driver will default to not using encryption.
I'm not entirely comfortable with this, but doing encrypted websockets
in the browser is a massive hurdle for users, since each Neo4j database
has it's own self-signed certificate by default. It seems that if we
forced first-time users to install new CA certificates in their
browsers, they will just opt to turn encryption off.