Better error handling for unpackable types
Whenever we encounter an unknown type we throw an exception midstream,
when we have already started writing data back to the channel. Instead of
starting to write to the channel write away we create an function that we
are certain will be packable, and fail early if we cannot pack.

Fixes #91