Split routing logic into multiple components
Previously it was mostly located in routing driver which owned the cluster
view and all logic to call `getServers` and interpret it's output/failures.

This introduces two new internal components:
 * `GetServersUtil` - responsible for calling `getServers` procedure and
 processing it's output
 * `Rediscovery` - responsible for fetching a routing table from the specified
 server

Routing driver now owns a routing table and an instance of rediscovery. It
uses the later one to update the routing table when needed.

Added dev dependency to lolex(https://github.com/sinonjs/lolex) to be able to
mock system time and test things better.