Close session and connection used for rediscovery
Previously session and the underlying connection were not returned to the pool
after the rediscovery procedure. This happened because rediscovery session was
not initalized with an appropriate callback.

This commit fixes the problem by extracting the release callback code into a
function that gets called for both "normal" and "rediscovery" sessions.
Probably a cleaner way to fix this would be to move functionality into
`Session#close()` and make connection release itself to the pool. However
this resulted into lots of non-trivial changes and thus should go into a
separate PR.