Fixed purging of connections after rediscovery
After rediscovery newly obtained routing table is compared to the known
routing table to find all servers that are "stale" - in the current table
but not in the new one. Connections to such servers are supposed to
be terminated.

However this did not happen because purging function was not invoked
correctly. This commit fixes the issue.