Extract routing table handling logic in a class
This commit introduces a `ConnectionProvider` abstraction that is mainly
responsible for acquiring connections from the connection pool.
Implementation for the direct driver does only that while implementation
for routing driver (`LoadBalancer`) handles routing table refreshes
and rediscovery. So all this logic was moved to the `LoadBalancer` class.

This refactoring allows easier testing and is the groundwork to decouple
session from connection. Idea is that sessions will hold `ConnectionProvider`
and use it to obtain connections when necessary.