Lazy connection initialization and eager release
This commit effectively decouples session from connection. Session now
holds two `ConnectionHolder`s that lazily pick connection from the pool
when it is needed and try to release connection back to the pool when it
does not have any users. Session keeps one connection holder for read
connection and one for write connection.

Connection is acquired for `Session#run()` and released when the
returned result is either fully consumed or error happens. Also
connection is acquired for `Session#beginTransaction()` and released
when transaction commits or rolls back.