Allow bookmark in `Driver#session()`
This commit adds second parameter `bookmark` to the existing
`Driver#session(accessMode)` API function to allow creation of a session
with the specified initial bookmark. This means that first transaction in the
created session will ensure that the server, it is connected to, is at
least up to the database version denoted by the bookmark.

Bookmark is a session related concept and this commit adds symmetry with
regards to `Session#lastBookmark()` method.