Fix streaming of queued records in stream observer
`StreamObserver` memorizes all incoming records when it does not have a
subscriber to forward records to. Later when such subscriber appears
it receives all queued records in on go.

This functionality did not work properly because of a missing `this`
which resulted in access to an undefined variable. This commit fixes
the problem and adds some tests.