Fix purging of connections on INIT failure
Commit moves waiting for connection initialization promise higher
in the stack - from `ConnectionProvider` to `ConnectionHolder`.
This is needed to be able to notify stream observer about used
connection even if initialization fails. It is later possible to
purge connections for the given server. Previously stream observer
did not get to know the connection and pool was not purged on
connectivity errors.