Multiple bookmarks support
Previously it was possible to only supply a single bookmark when creating a
new session. However there is a use-case to be able to supply multiple
bookmarks when multiple concurrent tasks execute write queries and then
reader should be able to observe all those writes. To achieve this
driver now allows passing an array of bookmarks to `Driver#session()`
function.

Driver will now send:
```
{
    bookmark: "max",
    bookmarks: ["one", "two", "max"]
}
```
instead of simple:
```
{
    bookmark: "max"
}
```

this is done to maintain backwards compatibility with databases that only
support a single bookmark. It forces driver to parse and compare bookmarks
which violates the fact that bookmarks are opaque. This is done only to
maintain backwards compatibility and should not be copied. Code doing
this will eventually be removed.

Related Bolt server PR: https://github.com/neo4j/neo4j/pull/9404