Extract load balancing logic in a separate class
Previously load balancing logic for read and write servers was coupled
with storage of server addresses in `RoundRobinArray` class. Such
approach was problematic because it made hard to use different load
balancing strategy.

This commit pulls load balancing in a separate class and removes
`RoundRobinArray`. Instead addresses are stored in simple arrays
and load balancing logic is applied to those arrays. Such approach
makes it easy to plug a different load balancing strategy.