Refactor boltstub, always run related tests
Previously driver allowed boltkit to be an optional dependency and
relied on neokit (submodule) to start/stop neo4j database. That's
why all tests that use stub server from boltkit were executed as a
separate step. These days boltkit is always required and
neokit is gone.

This commit refactors stub server handling and turns all boltkit tests
into normal tests that run as part of `npm test` and `runTests` script.
These tests are however skipped in browser testing because it is not
possible to spawn child processes when in phantomjs environment.