Support in-dev Neo4j version
Driver receives Neo4j version in response for INIT message. Released
versions have format `Neo4j/X.Y.Z` and in-dev version has `Neo4j/dev`.
Previously driver was not able to parse the later and failed.

This commit fixes the problem, `ServerVersion` will now be able to
represent an in-dev version. Such `ServerVersion` is treated lower
than any other version because driver can't know which features it
supports and which it does not.