Make IPv6 WebSocket work in Edge
WebSocket in IE and Edge does not understand IPv6 addresses because they
contain ':'. Which is an invalid symbol for UNC
https://en.wikipedia.org/wiki/IPv6_address#Literal_IPv6_addresses_in_UNC_path_names.
This caused `new WebSocket()` call to throw `SyntaxError` for all given
IPv6 addresses.

Following URL rewriting is needed to make WebSocket work with IPv6:
  1) replace all ':' with '-'
  2) replace '%' with 's' for link-local addresses
  3) append magic '.ipv6-literal.net' suffix

After such changes WebSocket in IE and Edge can connect to an IPv6
address.

This commit makes `WebSocketChannel` catch `SyntaxError` and rewrite
IPv6 address. It chooses not to detect browser/OS and things like this
upfront because it seems hard. User-agent string changes often and
differs for different versions of IE.

See following links for more details:
https://social.msdn.microsoft.com/Forums/ie/en-US/06cca73b-63c2-4bf9-899b-b229c50449ff/whether-ie10-websocket-support-ipv6?forum=iewebdevelopment
https://www.itdojo.com/ipv6-addresses-and-unc-path-names-overcoming-illegal/
https://en.wikipedia.org/wiki/IPv6_address#Literal_IPv6_addresses_in_UNC_path_names