Fix parsing of connectionTimeout setting
Connection timeout limits amount of time connection spends waiting for
the remote side to respond. It makes connect not hang forever if remote
side does not respond. Configured positive values define the actual
timeout. Zero and negative values mean infinite timeout.

Previously configured value was not correctly interpreted and thus it
was not possible to disable the timeout. Zero and negative values
resulted in a default value of 5 seconds to be used.

This commit fixes the problem so that zero and negative values mean
no timeout.