Clear WebSocket connection timeout when closing channel
`WebSocketChannel` is built on top of a `WebSocket` and contains
property that references it. `WebSocket` connection timeout is
enforced by a separate setTimeout timer that closes the socket
after configured amount of time. Previously this timeout has
only been cleared when connection is established. It has not
been cleared when channel is closed, resulting in potential
existence of stray timers.

This commit makes `WebSocketChannel` clear the connection timeout timer
when it's closed. So connect timeout timer is removed even if channel
is closed before connection is established.