Support for 2D and 3D points in Bolt V2
This commit makes driver able to send and receive 2D and 3D points.
Both are represented as custom `Point2D` and `Point3D` classes that
contain coordinates and coordinate reference system identifier. This
identifier comes as an integer from the database. It can thus be either
represented as `Integer` or native `number`, when
`disableLosslessIntegers` config option is set to true.