More validation for query parameters
Driver expects query parameters to either be undefined/null or an
object. This was previously not enforced and illegal parameters,
like strings or arrays, were sent to the database. This resulted
in a protocol violation and database closed the connection. Users
were only able to see the actual error/stacktrace in the database
logs. Driver received a `ServiceUnavailable` or `SessionExpired`
error.

This commit adds validation of query parameters. It also prohibits
nodes, relationships and paths from being used as query parameters
in the driver.