Fix release of broken connections to the pool
An attempt to resolve a pending acquisition request can be made when
connection is returned back to the pool. It will only be made when
there exists a pending acquisition request and returned connection
is valid.

It was previously possible for the release operation to deque a pending
acquisition request and forget about it, even when the returned
connection was invalid. This made the acquisition request either
fail with acquisition timeout or with "Cannot read property 'filter'
of undefined" error. Later happened because subsequent release operation
for the same key removed the array of pending acquisition requests.

This commit fixes the problem by making release operation only deque an
acquisition request when new request can be acquired.