Improve release of broken connection when pool is full
Release operation discards the connection if it is broken. It also tries
to resolve a pending acquisition request, if any. In order to this it
tries to execute a promise-free internal acquire that immediately
returns an available resource or creates a new one, if possible. New
resource can only be created if pool is not at its maximum capacity.

Previously, release operation only decremented the active connection
counter at the end. This made it impossible to create a new connection
when broken one is returned.

This commit fixes the problem by making `Pool#_release()` decrement
the active counter after connection has been either added to the list
of idle connections or discarded because it turned out to be broken.