Update to babel-preset-env
Previously used presets 'es2015' and 'stage-3' are now deprecated in
favour of 'env'. This resulted in a build warning. New preset makes it
possible to specify target browsers and NodeJS versions. Only needed
code transformations and polyfills are then applied. Right now browsers
and NodeJS versions are not configured so this preset behaves the same
was as the previous two.

Also updated versions of babel-related dependencies.

For more details about the 'env' preset see:
http://babeljs.io/env
https://github.com/babel/babel/tree/master/packages/babel-preset-env