Warn when driver config inconsistent with web page protocol
Encryption can be explicitly turned on or off in the driver config.
Driver can be used on both HTTP and HTTPS web pages. Configured
encryption has to match the security protocol of the web page. Insecure
WebSockets ('ws') have to be used on insecure pages ('http'). Secure
WebSockets ('wss') have to be used on secure pages ('https').
WebSockets might not work in a mixed content environment,
i.e. 'ws' + 'https' or 'wss' + 'http'.

This commit makes driver print a warning when its configuration is
inconsistent with the web page.