Improve failure handling for RESET and ACK_FAILURE
Both RESET and ACK_FAILURE messages will now be immediately flushed to
the network to make sure connections are fully cleaned up before being
returned to the pool. Outbound messages will only be sent if the
connection is not broken. Failure to RESET or ACK_FAILURE will be
considered a fatal error, unless ACK_FAILURE was ignored because of
a subsequent RESET.