Upgrade karma to get rid of ES6 and ES7 polyfill
Previous Karma version pulled in a polyfill that made it possible to
use ES6 and ES7 features in tests. Such features might not be available
during runtime and could result in errors. `String#padStart()` is an
example of a feature that was used in the code and was added by the
Karma polyfill. Tests worked just fine but driver code failed at
runtime in older NodeJS versions.

Note that Babel runtime that is used by the driver does not polyfill
instance methods such as `String#padStart()`.