Avoid forgetting address twice on error in routing driver
Routing driver removes address from the routing table on error. This
removal is triggered by a `StreamObserver` that observes the incoming
messages. Same observer is used for RUN and PULL_ALL. It used to
trigger the forget operation trice on every error. This wasn't harmful
but looked weird in logs.

This commit fixes the problem by making `StreamObserver` invoke the
provided callback only once.