Fix conversion from standard Date to Time & DateTime
Time zone offset was not converted correctly. Offset exposed by the
standard JavaScript `Date` is the difference, in minutes, from local
time to UTC. So positive offset means local time is behind UTC and
negative means it's ahead. This is different from Neo4j temporal types
that support time zone offsets - `Time` and `DateTime`. They define
offset as the difference, in seconds, from UTC to local time. Previous
code converted offset in minutes to offset in seconds but did not
change the sign of the value.