Update uri-js
The new version of `uri-js` is not extracting the query params when the protocol is `ws` or `wss`, the query is being treated as
`resourceName` together with the resource path.

In maner to update this library version, a small patch was applied to handle this especial case.

(See https://github.com/garycourt/uri-js#wswss-support)