Supporting browser, integration and stress tests on testkit
Teskit doesn't run Neo4j on the localhost and with the same credentials of boltkit, but it exposes
the host and credentials as env variables.

So all this configuration being loaded to karma on the config files, parsed in `shared-neo4j` with fallback to
the process env (case it's running the usual integration test) and to the default boltkit values.

All the integration tests were changed to use the values from the shareNeo4j object.

Disabled tests:
* IPV6 tests are skipped on the testkit because of the lack of support for this
* TLS tests are disabled since it's already tested on the testkit suite
* Tests which depends on the enterprise edition of Neo4j is running only on this scenario