Refactoring routing table rediscovery routine
The goal of this change is prepare the terrain for loading the routing table using the routing message.

Previously, the `Rediscovery` class was responsible for the parsing the result for the procedure applying the business rules and the `RoutingUtil` was responsible for calling the correct procedure depending on the protocol version and provide methods to parse the result.

In the new design,

* `Rediscovery` is responsible for orchestrating the `Session`, `RoutingTableGetterFactory` and `RoutingTableGetter` to get the `Table`.
* `RoutingTableGetterFactory` is responsible for creating the correct `RoutingTableGetter` configuration due the protocol version.
* `ProcedureRoutingTableGetter` is responsible for getting a valid `RoutingTable` running the configured procedure
* `SingleDatabaseProcedureRunner` is responsible for running the procedure used for single database protocol and return the result
* `MultiDatabaseProcedureRunner` is responsible for running the procedure used for multi database protocol and return the result

To extend this model with a new RoutingTable source, you need to create a new getter which will have the same blackbox behaviour of `ProcedureRoutingTableGetter` and instantiate it properly on the factory.