Create 4.3 Protocol and rediscovery the Routing Table using the Route Message (#646)
Starting the development of the version 4.3 with the creation of the protocol class making it available on the handshake. The replacement of the routing procedure by the route message for routing table discovery on this new version pushes the complexity to the server and makes the protocol cleaner and more efficient.

The routing table requisition details was pushed to the protocol implementation classes, the request provides a observer which parses the response and make it available as a RawRoutingTable on completed.

The rediscovery class uses the method, constructs the RoutingTable and treats the exceptional cases.