Improve connection/protocol layers
The implementation of the connection channel, the protocol and handshake are dependent in the other in cross mutual dependency situation. This cross dependency situation could leave to unexpected behaviours when something change on this layers.

The solution for this issue was separate the bolt protocol with it handshake process, creation, observers and so on in a different package to hide the details and make easier to spot the unwanted dependencies.

The following changes were done to achieve these goals:
* [x] Extract a factory for the connection removing the create method
* [x] Extract the handshake process and separate in two, the handshake message exchange and the protocol initialisation
* [x] Remove the response parsing and response observers management from the connection and put this in the bolt package (`response-handler.js`)
* [x] Move the message write responsibility to the protocol object
* [x] Remove connection dependency from the stream-observers
* [x] Fix the connection integration tests
* [x] Improve the unit tests
* [x] Do not touch or break test kit tests and stress tests