Providing protocol version in Result get by Transaction.run
The result summary was not filled with the protocol version because the transactions were creating
the Result object with a empty connection holder. Since the Result releases the connection from holder
when it completes, the original holder could not be passed to it.

The ReadOnlyConnectionHolder provides a safe way to pass the holder forward without let the consumer initialize,
close or release the connection by simulating this methods without do the action and delegating part of the requests to the original ConnectionHolder.

These changes also made shared-neo4j cluster friendly.