Run stress tests out of Jasmine (#667)
The goal of this change is improve the stability of the stress test avoiding unexpected timeouts. This change also provide a way to configure the stress-test to run for a defined time period, it's particular interesting for when it's running in test-kit in cluster mode simulating the change on the cluster composition.