Extracting a core module from the driver
The segmentation of the core module is part of the effort of reduce the size of the bundle for web application. On the core module, all the driver async interface will be defined and implemented in typescript, the details of the bolt protocol, pooling and packstream is not part of this core module.

The core package will be bundled together with `neo4j-driver`.