testkit: Update NodeJS version in the Dockerfile
The support for the `NodeJS 8.x` was dropped and the `npm update` was failing because of that.
The NodeJS updagrade is being done to `10.x` since it's the oldest version in maintenance.
This update was done by updating the Ubuntu version since the other of updates was broken the TLS configuration. Along with this, the TLS 1.0 was enable in the OpenSSL config since it's disable by default in this ubuntu's version.