Fix connection close by server treatment by websocket channels
The lack of set channel._open to false when the onclose event is triggered was causing the channel be broken without the other parts of the driver notice. In tbis way, a next iteration trying to get the broken connection will succeded in the try and run the query will result in a eternal pending promise.

Mark the channel as closed enable the pool to discard and create a new connection if it needed.