Extracting Bolt Connection Module (#1)
The segmentation of the bolt-connection module is part of the effort of reduce the size of the bundle for web application. On the bolt-connection module, all the connection and connection provider details are implemented. Chunking, Pool, Routing, Bolt Protocol, etc.

This segmentation is part of the effort to modularise better the driver.