Add support for BigInt
Recent V8 release adds a new type - BigInt. It is thus available in latest NodeJS (v10+) and Chrome (v67+).

Driver could use this type instead of the custom Integer to represent lossy integer numbers. This is configurable because BigInt and Number are not fully compatible, so driver can't always return BitInts. For example `BigInt(1) / Number(1)` results in a TypeError, as well as `JSON.stringify({v: BigInt(1)})`.

The configuration property is `useBigInt` and it's `false` by default, in manner to break drivers user on update.

This configuration take precedence over `disableLosslessIntegers`