Restructuration of Driver in Modules  (#678)
The segmentation of the core module is part of the effort of reduce the size of the bundle for web application.

On the core module, all the driver async interface will be defined and implemented in typescript, the details of the bolt protocol, pooling and packstream is not part of this core module.

bolt-connection module is responsible for implementing the connectivity details. The bolt protocol, handshake, packstream, connection pool, browser-channel and node-channel are defined in this module. It will provide a way to bundle the correct channel for node and browser using the property browser on the package.json.

The core and bolt-connection package will be bundled together with neo4j-driver.