Fix log message doesn't appearing correctly in the Bolt log.
The write and receive message functions were moved from the connection to the `response-handler` and `bolt-protocol`, but logs on the functions still creating the message using the `this`, but in the nnew context the value of `this` is not serialized as the connection id.

Since the connection id data is dynamic and the depends on a response from the server, the issue was solved decorating the logger function to add connection context information.