Introducion Bearer tokens and SSO support
This change provides an utility function to easy create auth token object for `bearer` tokens.

The error handling related with this kind of token usage is also implemented. When a `Neo.ClientError.Security.TokenExpired` error happens, the adress should be purged from the pool and the caller should be notified - no retry should be performed.