Introducing Bearer tokens and SSO support (#783)
This change provides a utility function for easy create auth token object for `bearer` tokens.

The error handling related with this kind of token usage is also implemented. When a `Neo.ClientError.Security.TokenExpired` error happens, the address should be purged from the pool and the caller should be notified - no retry should be performed.