Make test script handle runs as root (#787)
The NPM lifecycle scripts can not be run by the root user for security reasons. The full build of the driver repo depends on the `prepare` script run during the build (or after it).

The solution for this case is run the prepare script if the current user is the root, the way the driver package will be ready for testing.