Improve performance for packstream in browser.
The buffer and utf-8 encoding implemetation in browser was using the heap and `text-encoding-utf-8` library respectively. This process was really slower than in nodejs. For example, a string with 85000 length was taking 180 seconds to serialize in the browser, the same string takes miliseconds to be serialized in nodejs.

This performance issue was resolved by using the libraries `buffer` and `string_decoder`, libraries which is already available in nodejs, also in browser. This substitution bring the serialization of the 85000 length string to take only 350 ms when it's done in browser.

The `text-encoding-utf-8` was removed from the project dependency in the process.