Fixing summary consumption.
Summary was not available with the correct information after the result being consumed. This fix, solve this issue by wrap the summary with the same treatment subscribe and then have on complete.

This change also introduce a cache for capturing the last summary the stream complete, this way the summary function could be called after the results get consumed.

PS: The same fixes were applied to keys()