testkit: Fix `Unknown request` error path (#845)
There was two different errors during the unknown request error treatment.

First, the stringify function was not define in the LocalController scope. It get solve by extract this function to it own file and import it in the local.js.

Second, the backend was capturing the error in the `request` event and calling `writeBackendError` with the error object instead of the message.