AsyncIterator API for Result consumption (#831)
The current driver implementation has only two options for record consumption, the default async/promise api and the reactive api. 

In the async side, there is no simple way to interact over the results before consuming all the data stream and back-pressure the consumption. Meanwhile in the reactive, the back-pressure is implemented  but this api usage depends on the knowledge of reactive programming and `RxJS`.  Another issue with our current reactive implementation is not having the way to configure the back-pressure or pause the stream consumption. 

The introduction of the Async Iterator is part of a group of improvements to make the driver more idiomatic and memory efficient.