Introduce `.peek()` to the Async Iterator API
This feature allows the user peek the element on the head of the cursor without moving the cursor forward.

The implementation of this feature has as side effect a new implementation `.return` method in the iterator.
This new version of the method does most the same thing as the previous one, but it also cancels the stream
since this method moves the cursor to the end.

The `next` method also get tweaked.